pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-dedup"
version = "0.1.0"
description = "Detect duplicate and similar images using SHA256 and perceptual hashing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Fernando Pelliccioni", email = "fpelliccioni@gmail.com" }
]
keywords = ["image", "duplicate", "deduplication", "hash", "perceptual-hash"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "Pillow>=10.0.0",
    "imagehash>=4.3.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
image-dedup = "image_dedup.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/image_dedup"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
